added e2e test for operator resolution
